\name{pr.ahrq.ICD9.5digit}
\alias{pr.ahrq.ICD9.5digit}
\title{Converts an icd9 code to numeric format}
\usage{
pr.ahrq.ICD9.5digit(icd.code)
}
\arguments{
  \item{icd.code}{A string numeric code}
}
\value{
\code{float} Returns a decimal code XXX.XX
}
\description{
Converts an icd9 code to numeric format
}
\seealso{
\code{\link{ahrq}}
}

