% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pr.ahrq.list}
\alias{pr.ahrq.list}
\title{Gets the core AHRQ list}
\usage{
pr.ahrq.list(version)
}
\arguments{
  \item{version}{The version number of interest}
}
\value{
\code{list} A full AHRQ-list with all the numbers of
interest
}
\description{
The function gets the list, uses the 3.5 version and
updates according to the codes specified in the different
versions as presented in "Changes to the Comorbidity
Software" on the
\url{http://www.hcup-us.ahrq.gov/toolssoftware/comorbidity/comorbidity.jsp}
webpage.
}
\keyword{internal}

