% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pr.ahrq.preprocess.v.codes}
\alias{pr.ahrq.preprocess.v.codes}
\title{Converts icd9 codes with V at the first letter into numeric format}
\usage{
pr.ahrq.preprocess.v.codes(v.code)
}
\arguments{
  \item{v.code}{A string numeric icd-code that starts with
  a V}
}
\value{
\code{int} Returns an int with 5 digits
}
\description{
Converts icd9 codes with V at the first letter into numeric
format
}
\seealso{
\code{\link{ahrq}}
}
\keyword{internal}

