\name{pr.elixhauser.ICD9.5digit}
\alias{pr.elixhauser.ICD9.5digit}
\title{Convert to 5 digit ICD-code}
\usage{
pr.elixhauser.ICD9.5digit(codes)
}
\arguments{
  \item{codes}{\code{string/vector} indicating the code}
}
\value{
\code{integer} Returns a integer value XXXXX
}
\description{
Convert to 5 digit ICD-code
}
\seealso{
\code{\link{elixhauser}},
\code{\link{cmrbdt.finder.numeric.elixhauser_Elixhauser1998}}
}

