% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pr.elixhauser.points.30}
\alias{pr.elixhauser.points.30}
\title{Score points for Elixhauser's comorbidity count}
\usage{
pr.elixhauser.points.30(input.frame)
}
\arguments{
  \item{input.frame}{This is a data frame with 5 digit
  ICD-9-CM codes as XXX.XX This can also be provided as a
  vector or matrix.}
}
\value{
\code{matrix} Returns a matrix with the one column per
comorbidity. No comorbidity is indicated by 0 while 1
indicates existing comorbidity.
}
\description{
The following function develops a matrix with rows devoted
to respondents and each column a comorbidity.
}
\seealso{
\code{\link{elixhauser}}
}
\keyword{internal}

