\name{prAHRQ.process.v.codes}
\alias{prAHRQ.process.v.codes}
\title{Converts icd9 codes with V at the first letter into numeric format}
\usage{
prAHRQ.process.v.codes(v.code)
}
\arguments{
  \item{icd.code}{A string numeric icd-code that starts
  with a V}
}
\value{
\code{int} Returns an int with 5 digits
}
\description{
Converts icd9 codes with V at the first letter into numeric
format
}
\seealso{
\code{\link{ahrq}}
}

