\name{prDeyo.ICD9.5digit}
\alias{prDeyo.ICD9.5digit}
\title{Convert to 5 digit ICD-code}
\usage{
prDeyo.ICD9.5digit(icd.code)
}
\arguments{
  \item{icd.code}{\code{string} indicating the code}
}
\value{
\code{float} Returns a float value XXX.XX
}
\description{
Convert to 5 digit ICD-code
}
\seealso{
\code{\link{deyo}
}

