% Generated by roxygen2 (4.0.2): do not edit by hand
\name{preproc.code.splitter}
\alias{preproc.code.splitter}
\title{Code splitter}
\usage{
preproc.code.splitter(icd, icd_ver, split_str = " ", trim = TRUE)
}
\arguments{
  \item{icd}{The ICD-codes to be stripped}

  \item{icd_ver}{Currently not used - this parameter is
  only for compatibility reasons as other preprocessing
  functions may require the knowledge of actual icd
  version.}

  \item{split_str}{The string that is the splitter}

  \item{trim}{If the string has trailing spaces these
  should usualy be removed}
}
\value{
\code{vector} A vector with only one code per entry. The
attr(,"icd_ver") now contains the corresponding
icd_version.
}
\description{
As the number of codes per admission frequently may differs
between visits and patients it can be efficient to store
them in a string that is separated by a " " character os
something similar. In case you want something other than a
space you need to encapsulate the function.
}
\examples{
preproc.code.splitter(icd=c("M161", "M161 ", "M161 J445"),
                  icd_ver=c(10, 10, 10))
}
\seealso{
Other preprocessor functions:
\code{\link{preproc.Swedich.ICD9}};
\code{\link{preproc.strip.dot}}
}

