\name{prGetOutVector}
\alias{prGetOutVector}
\title{A internal function for generating an output vector containing the score elements}
\usage{
prGetOutVector(out, score)
}
\arguments{
  \item{out}{A vector or NULL}

  \item{score}{The score list with element names}
}
\value{
\code{vector} A named vector
}
\description{
Throws errors if the output has an invalid format
}

