% Generated by roxygen2 (4.0.0): do not edit by hand
\name{elixhauser}
\alias{elixhauser}
\title{Produces an Elixhauser comorbidity count}
\usage{
elixhauser(input.frame)
}
\arguments{
  \item{input.frame}{This is a data frame with 5 character
  ICD-9-CM codes without decimal points}
}
\value{
A list of one vector and one data frames
\item{COMORBIDITY.CT}{A vector containing the number of
comorbidtities for each patient} \item{COMORBIDITIES}{A
data frame denoting which diagnoses each patient has with a
1}
}
\description{
This function take as input a data frame structured such
that each row contains a list of ICD-9-CM codes attributed
to a single patient. The function goes from row to row
comparing the ICD-9-CM codes a patient has with the
Elixhauser comorbidity index diagnosis codes.  If a patient
has a diagnosis (as indicated by ICD-9-CM code) that
belongs to one of the Elixhauser diagnoses, then the
patient is considered to have this diagnosis. Regardless of
how many different ICD-9-CM codes a patient has
corresponding to a particular comorbidity category, a
comorbidity is only counted once.
}
\details{
The value returned consists of a vector and a data frames.
The vector is the total comorbidity count.  In the data
frame returned each row in the data frame is devoted to a
particular patient, and each column is a diagnosis. The
data frame codes a 0 if the patient does not have that
diagnosis and 1 if the patient does have that diagnosis.

The ICD-9-CM codes must be 5 characters long and have no
decimal poinst.If any codes are less than 5 digits, then
they must be led by zeroes or followed by the letter "X" to
create a 5 character code.  No more than 2 "X" ' s can be
used at the tail end, or a meaningful ICD-9-CM code cannot
be derived.  If your ICD-9-CM codes contain decimals or
some other filler character, then they must be converted to
the format above for this function to work properly.
}
\section{Warning}{
  This function is deprecated. We recommend you switch to
  the more general \code{\link{cmrbdt.calc}} function
  instead.
}
\examples{
x <- matrix(NA, nrow = 3, ncol = 2)
x[1,] <- c("41000", "42800")
x[2,] <- c("57220", "1961X")
x[3,1] <- "042XX"
x <- as.data.frame(x)
elixhauser(x)
}
\author{
Paul Gerrard, Max Gordon
}
\references{
Elixhauser A, Steiner C, Harris DR, Coffey RM. (1998)
Comorbidity measures for use with administrative data.  Med
Care. 36:8-27.
}
\keyword{internal}

