% Generated by roxygen2 (4.0.2): do not edit by hand
\name{weight.Charlsons.org}
\alias{weight.Charlsons.org}
\title{Score points for Charlson's original comorbidity weights}
\usage{
weight.Charlsons.org(ds)
}
\arguments{
  \item{ds}{This is a \code{data.frame} or \code{matrix}
  with the comorbidity groups as returned by any of the
  Charlson functions}
}
\value{
Weighted comorbidities
}
\description{
The following function adds the weights of the Charlson
original weighting system where comorbidities were score
between 1 and 6.
}
\examples{
out <- cmrbdt.finder.regex.charlson_Quan2005("I252")
weight.Charlsons.org(out)
}
\seealso{
Other weight functions:
\code{\link{weight.Charlsons.Quan2011}};
\code{\link{weight.Elixhausers.VanWalraven2009}}
}

